Fix ROUND() cast in diag endpoint for PostgreSQL
ROUND(double precision, integer) requires ::numeric cast in PostgreSQL.

+++ backend/server.js
@@ -1005,7 +1005,7 @@ app.get('/api/diag/queries', async (req, res) => {
       pool.query(`
         SELECT p.name AS account, COUNT(*) AS total_closed,
           SUM(CASE WHEN t.exit_reason = 'WINDOW_CLOSE'    THEN 1 ELSE 0 END) AS window_close_count,
-          ROUND(SUM(CASE WHEN t.exit_reason = 'WINDOW_CLOSE' THEN 1 ELSE 0 END)*100.0/COUNT(*),1) AS window_close_pct,
+          ROUND((SUM(CASE WHEN t.exit_reason = 'WINDOW_CLOSE' THEN 1 ELSE 0 END)*100.0/COUNT(*))::numeric,1) AS window_close_pct,
           SUM(CASE WHEN t.exit_reason = 'TARGET_HIT'      THEN 1 ELSE 0 END) AS target_hit,
           SUM(CASE WHEN t.exit_reason = 'STOP_HIT'        THEN 1 ELSE 0 END) AS stop_hit,
           SUM(CASE WHEN t.exit_reason = 'CIRCUIT_BREAKER' THEN 1 ELSE 0 END) AS circuit_breaker,
@@ -1015,7 +1015,7 @@ app.get('/api/diag/queries', async (req, res) => {
         WHERE t.exit_reason IS NOT NULL GROUP BY p.name, p.id ORDER BY p.id`),
       pool.query(`
         SELECT p.name AS account, p.circuit_breaker_date, p.day_start_balance, p.current_balance,
-          ROUND(p.current_balance - p.day_start_balance, 2) AS day_pnl
+          ROUND((p.current_balance - p.day_start_balance)::numeric, 2) AS day_pnl
         FROM portfolios p WHERE p.circuit_breaker_date IS NOT NULL ORDER BY p.id`),
       pool.query(`
         SELECT p.name AS account, COUNT(*) AS total_cycles_with_open_positions,
@@ -1030,7 +1030,7 @@ app.get('/api/diag/queries', async (req, res) => {
       pool.query(`
         SELECT p.name AS account, COUNT(*) AS total_trades,
           SUM(CASE WHEN t.exit_reason = 'NO_ENTRY' THEN 1 ELSE 0 END) AS no_entry_count,
-          ROUND(SUM(CASE WHEN t.exit_reason = 'NO_ENTRY' THEN 1 ELSE 0 END)*100.0/COUNT(*),1) AS no_entry_pct,
+          ROUND((SUM(CASE WHEN t.exit_reason = 'NO_ENTRY' THEN 1 ELSE 0 END)*100.0/COUNT(*))::numeric,1) AS no_entry_pct,
           SUM(CASE WHEN t.exit_reason = 'EXPIRED' THEN 1 ELSE 0 END) AS expired_count
         FROM trades t JOIN portfolios p ON p.id = t.portfolio_id
         WHERE t.exit_reason IS NOT NULL GROUP BY p.name, p.id ORDER BY p.id`),
